+++ src/libsac2c/tree/tree_basic.c
@@ -149,7 +149,7 @@ TBmakeAccess (node *array, node *iv, accessclass_t mclass, shpseg *offset,
 /*--------------------------------------------------------------------------*/
 
 rc_t *
-TBmakeReuseCandidate (node *array, int dim, rc_t *next)
+TBmakeReuseCandidate (node *array, size_t dim, rc_t *next)
 {
     rc_t *tmp;
     int i;
@@ -178,7 +178,7 @@ TBmakeReuseCandidate (node *array, int dim, rc_t *next)
 /*--------------------------------------------------------------------------*/
 
 cuda_index_t *
-TBmakeCudaIndex (unsigned int type, int coefficient, node *id, int looplevel,
+TBmakeCudaIndex (unsigned int type, int coefficient, node *id, size_t looplevel,
                  cuda_index_t *next)
 {
     cuda_index_t *idx;
